+++ examples/echo/src/main.rs
@@ -192,7 +192,7 @@ async fn list(bus: &Handle) -> Result<()> {
     // services. Every call to `add_object` registers interest in one type of object.
     let mut discoverer = bus
         .create_discoverer()
-        .any((), [EchoProxy::UUID])
+        .any_object_with_services((), [EchoProxy::UUID])
         .build_current_only()
         .await?;